Minor modifications. Audio scraper tool
Modified some minor things, such as the name of the API database.
It used to be KanjiKana.sqlite, but since this database is also going to be used to store client info, it was renamed to simply API.sqlite.

Additionally, we now have a new tool Blackbeard that scrapes audio from a certain webste. Sneaky.